Create example1_php.php
Small example of how to connect to Grupo Incofisa's WebService using PHP 7.X.

Note that the class maintains the session, therefore it avoid validate the user on each request. This is very important to get the fastest possible response. If it were validated in each request, it would add several seconds to the operation apart from the fact that the system could block your user due to excessive logins.

Do not forget to put your login data in the script, there are two defines for it: USERNAME and PASS.

We are not regular PHP programmers so any improvement is welcome